requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiogram_i18n==1.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asyncpg==0.29.0
attrs==24.1.0
Babel==2.15.0
base58==2.1.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
distro==1.9.0
fluent.runtime==0.4.0
fluent.syntax==0.19.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
netaddr==1.3.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
yookassa==3.3.0
